requirements.txt
# Google Generative AI SDK (latest)
google-generativeai>=0.8.0

# LangChain and LangGraph for advanced AI workflows
langchain>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# Environment management
python-dotenv>=1.0.0

# FastAPI framework (latest)
fastapi>=0.115.0

# ASGI server with performance improvements
uvicorn[standard]>=0.32.0

# Additional dependencies for production
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Testing and security
colorama>=0.4.6  # For colored terminal output in tests

# Security and CORS
python-multipart>=0.0.18

# MySQL Database
mysql-connector-python>=9.1.0
SQLAlchemy>=2.0.36

# Authentication
PyJWT>=2.10.1
passlib>=1.7.4
bcrypt>=4.2.1
